before making interactions

+++ tarzanproject/src/map/MapGenerator.java
@@ -1,5 +1,5 @@
 package map;
 
 public interface MapGenerator {
-	public double[][] createMap(int size); // assume square map
+	public float[][] createMap(int size); // assume square map
 }
